Missed updating the prototype for makeBlankMask
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1001 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/param.hh
@@ -150,7 +150,7 @@ namespace duchamp
     bool   isStatOK(int x, int y, int z);
 
     /// @brief Make a mask array -- an array saying whether each pixel is BLANK or not
-    bool  *makeBlankMask(float *array, int size);
+    bool  *makeBlankMask(float *array, size_t size);
 
     /// @brief Make a mask array -- an array saying whether each pixel is able to be used for statistics calculations
     bool *makeStatMask(float *array, size_t *dim);
